pyproject.toml
[tool.poetry]
name = "openapi-convert"
version = "0.1.0"
description = ""
authors = ["dewdroid <thossapoldew@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
markdown = "^3.4.3"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
